Conflicts:
	actor-sdk/sdk-core/core/core-shared/src/main/java/im/actor/core/modules/calls/CallMasterActor.java

+++ actor-server/actor-models/src/main/scala/im/actor/server/mtproto/protocol/ProtoMessage.scala
@@ -17,7 +17,7 @@ sealed trait IncomingProtoMessage
 sealed trait OutgoingProtoMessage
 
 sealed trait ResendableProtoMessage {
-  def bodySize: Int
+  def bodySize: Long
 }
 
 @SerialVersionUID(1L)
